feat practice-namespaces Finish namespace imports
