+++ src/components/project-list.ts
@@ -1,0 +1,68 @@
+/// <reference path="component.ts" />
+
+namespace App {
+  export class ProjectList extends Component<HTMLDivElement, HTMLElement> implements DragTarget {
+    assignedProjects: Project[];
+
+    constructor(private type: "active" | "finished") {
+      super('project-list', 'app', false, `${type}-projects` );
+      this.assignedProjects = [];
+
+
+      this.configure();
+      this.renderContent();
+    }
+
+    @autobind
+    dragOverHandler(event: DragEvent) {
+      if (event.dataTransfer && event.dataTransfer.types[0] === 'text/plain') {
+        event.preventDefault();
+        const listEl = this.element.querySelector('ul')!;
+        listEl.classList.add('droppable');
+      }
+    }
+
+    @autobind
+    dropHandler(event: DragEvent) {
+      const projectId = event.dataTransfer!.getData("text/plain");
+      projectState.moveProject(projectId, this.type === "active" ? ProjectStatus.Active : ProjectStatus.Finished);
+    }
+
+    @autobind
+    dragLeaveHandler(_: DragEvent) {
+      const listEl = this.element.querySelector('ul')!;
+      listEl.classList.remove('droppable');
+    }
+
+    private renderProjects() {
+      this.element.addEventListener('dragover', this.dragOverHandler);
+      this.element.addEventListener('dragleave', this.dragLeaveHandler);
+      this.element.addEventListener('drop', this.dropHandler);
+
+      const listEl = document.getElementById(`${this.type}-projects-list`)! as HTMLUListElement;
+      listEl.innerHTML = "";
+      for(const projectItem of this.assignedProjects) {
+        new ProjectItem(this.element.querySelector('ul')!.id, projectItem)
+      }
+    }
+
+    configure() {
+      projectState.addListener((projects: Project[]) => {
+        const relevantProjects = projects.filter(prj => {
+          if (this.type === 'active') {
+            return prj.status === ProjectStatus.Active
+          }
+          return prj.status === ProjectStatus.Finished
+        })
+        this.assignedProjects = relevantProjects;
+        this.renderProjects();
+      })
+    }
+
+    renderContent() {
+      const listId = `${this.type}-projects-list`;
+      this.element.querySelector("ul")!.id = listId;
+      this.element.querySelector("h2")!.textContent = this.type.toUpperCase() + " PROJECTS"
+    }
+  }
+}
