+++ src/decorators/autobind.ts
@@ -1,0 +1,13 @@
+namespace App {
+  export function autobind(_: any, _2: string, descriptor: PropertyDescriptor) {
+    const originalMethod = descriptor.value;
+    const adjDescriptor: PropertyDescriptor = {
+      configurable: true,
+      get() {
+        const boundFn = originalMethod.bind(this);
+        return boundFn;
+      }
+    }
+    return adjDescriptor;
+  }
+}
